fix #ACT-4304, ignore transient vars when deleting

+++ modules/activiti-engine/src/main/java/org/activiti/engine/impl/persistence/entity/ExecutionEntityManagerImpl.java
@@ -566,12 +566,14 @@ public class ExecutionEntityManagerImpl extends AbstractEntityManager<ExecutionE
         (enableExecutionRelationshipCounts && ((CountingExecutionEntity) executionEntity).getVariableCount() > 0)) {
       Collection<VariableInstance> executionVariables = executionEntity.getVariableInstancesLocal().values();
       for (VariableInstance variableInstance : executionVariables) {
-        VariableInstanceEntity variableInstanceEntity = (VariableInstanceEntity) variableInstance;
-        
-        VariableInstanceEntityManager variableInstanceEntityManager = getVariableInstanceEntityManager();
-        variableInstanceEntityManager.delete(variableInstanceEntity);
-        if (variableInstanceEntity.getByteArrayRef() != null && variableInstanceEntity.getByteArrayRef().getId() != null) {
-          getByteArrayEntityManager().deleteByteArrayById(variableInstanceEntity.getByteArrayRef().getId());
+        if (variableInstance instanceof VariableInstanceEntity) {
+          VariableInstanceEntity variableInstanceEntity = (VariableInstanceEntity) variableInstance;
+          
+          VariableInstanceEntityManager variableInstanceEntityManager = getVariableInstanceEntityManager();
+          variableInstanceEntityManager.delete(variableInstanceEntity);
+          if (variableInstanceEntity.getByteArrayRef() != null && variableInstanceEntity.getByteArrayRef().getId() != null) {
+            getByteArrayEntityManager().deleteByteArrayById(variableInstanceEntity.getByteArrayRef().getId());
+          }
         }
       }
     }
